Implement Light class. Change float to double.

+++ ray-tracing-from-scratch/main.cc
@@ -6,8 +6,8 @@
 int main() {
     std::cout << "\n=============STARTING!===========\n";
 
-    float image_height = 500.0f;
-    float image_width = 500.0f;
+    double image_height = 500.0f;
+    double image_width = 500.0f;
     Painter painter = Painter(
         (Canvas){image_width, image_height},
         /*file_name=*/"image.ppm",
@@ -34,7 +34,7 @@ int main() {
     Sphere sphere4;
     sphere4.name = "sphere4";
     sphere4.radius = 5000.0f;
-    sphere4.color = (Color){255, 255, 0};
+    sphere4.color = (Color){255, 255, 0}; // yellow
     sphere4.center = (Vec3){0.0f, -5001.0f, 0.0f};
 
     Scene scene;
